@@ -7,18 +7,17 @@ issue_tracker: https://github.com/fcodelabs/fcode_bloc/issues
 homepage: https://www.fcodelabs.com
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  firebase_core: ^0.7.0
-  cloud_firestore: ^0.16.0
-  rxdart: ^0.25.0
-  meta: ^1.1.8
+  firebase_core: ^1.0.0
+  cloud_firestore: ^1.0.0
+  rxdart: ^0.26.0
 
 dev_dependencies:
-  effective_dart: ^1.2.4
+  effective_dart: ^1.3.1
   flutter_test:
     sdk: flutter
 
